added delete cart items after payment

+++ payment/views.py
@@ -219,12 +219,13 @@ def process_order_verify(request):
     res_data = response.json()
     if res_data['status'] and res_data['data']['status'] == 'success':
         #clear cart
-        Cart.objects.filter(user=request.user).delete()
+        if request.user.is_authenticated:
+            Cart.objects.filter(user=request.user).delete()
         #clear session key
         if 'cart' in request.session:
             del request.session['cart']
             request.session.modified = True
-            
+
         return render(request,'payment/payment_success.html', {'details':res_data['data']})
     else :
         return render(request,'payment/payment_failed.html')
